requirements.txt
numpy
torch
torchvision
open_clip_torch
onnx
onnxsim
onnxruntime
onnxscript
opencv-python
tqdm
scipy
